@@ -3,7 +3,6 @@
 package step
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"os"
@@ -236,7 +235,7 @@ func TestCheckResourceIntegration(t *testing.T) {
 		t.Run(test.testName, func(t *testing.T) {
 			namespace := fmt.Sprintf("kuttl-test-%s", petname.Generate(2, "-"))
 
-			err := testenv.Client.Create(context.TODO(), kubernetes.NewResource("v1", "Namespace", namespace, ""))
+			err := testenv.Client.Create(t.Context(), kubernetes.NewResource("v1", "Namespace", namespace, ""))
 			if !k8serrors.IsAlreadyExists(err) {
 				// we are ignoring already exists here because in tests we by default use retry client so this can happen
 				assert.Nil(t, err)
@@ -245,7 +244,7 @@ func TestCheckResourceIntegration(t *testing.T) {
 				_, _, err := kubernetes.Namespaced(testenv.DiscoveryClient, actual, namespace)
 				assert.Nil(t, err)
 
-				assert.Nil(t, testenv.Client.Create(context.TODO(), actual))
+				assert.Nil(t, testenv.Client.Create(t.Context(), actual))
 			}
 
 			step := Step{
@@ -312,20 +311,20 @@ func TestStepDeleteExistingLabelMatch(t *testing.T) {
 
 	namespaceObj := kubernetes.NewResource("v1", "Namespace", namespace, "default")
 
-	assert.Nil(t, testenv.Client.Create(context.TODO(), namespaceObj))
-	assert.Nil(t, testenv.Client.Create(context.TODO(), podToKeep))
-	assert.Nil(t, testenv.Client.Create(context.TODO(), podToDelete))
-	assert.Nil(t, testenv.Client.Create(context.TODO(), podToDelete2))
+	assert.Nil(t, testenv.Client.Create(t.Context(), namespaceObj))
+	assert.Nil(t, testenv.Client.Create(t.Context(), podToKeep))
+	assert.Nil(t, testenv.Client.Create(t.Context(), podToDelete))
+	assert.Nil(t, testenv.Client.Create(t.Context(), podToDelete2))
 
-	assert.Nil(t, testenv.Client.Get(context.TODO(), kubernetes.ObjectKey(podToKeep), podToKeep))
-	assert.Nil(t, testenv.Client.Get(context.TODO(), kubernetes.ObjectKey(podToDelete), podToDelete))
-	assert.Nil(t, testenv.Client.Get(context.TODO(), kubernetes.ObjectKey(podToDelete2), podToDelete2))
+	assert.Nil(t, testenv.Client.Get(t.Context(), kubernetes.ObjectKey(podToKeep), podToKeep))
+	assert.Nil(t, testenv.Client.Get(t.Context(), kubernetes.ObjectKey(podToDelete), podToDelete))
+	assert.Nil(t, testenv.Client.Get(t.Context(), kubernetes.ObjectKey(podToDelete2), podToDelete2))
 
 	assert.Nil(t, step.DeleteExisting(namespace))
 
-	assert.Nil(t, testenv.Client.Get(context.TODO(), kubernetes.ObjectKey(podToKeep), podToKeep))
-	assert.True(t, k8serrors.IsNotFound(testenv.Client.Get(context.TODO(), kubernetes.ObjectKey(podToDelete), podToDelete)))
-	assert.True(t, k8serrors.IsNotFound(testenv.Client.Get(context.TODO(), kubernetes.ObjectKey(podToDelete2), podToDelete2)))
+	assert.Nil(t, testenv.Client.Get(t.Context(), kubernetes.ObjectKey(podToKeep), podToKeep))
+	assert.True(t, k8serrors.IsNotFound(testenv.Client.Get(t.Context(), kubernetes.ObjectKey(podToDelete), podToDelete)))
+	assert.True(t, k8serrors.IsNotFound(testenv.Client.Get(t.Context(), kubernetes.ObjectKey(podToDelete2), podToDelete2)))
 }
 
 func TestCheckedTypeAssertions(t *testing.T) {
@@ -348,10 +347,7 @@ func TestCheckedTypeAssertions(t *testing.T) {
 }
 
 func TestApplyExpansion(t *testing.T) {
-	os.Setenv("TEST_FOO", "test")
-	t.Cleanup(func() {
-		os.Unsetenv("TEST_FOO")
-	})
+	t.Setenv("TEST_FOO", "test")
 
 	step := Step{Dir: "test_data/assert_expand/"}
 	path := "test_data/assert_expand/00-step1.yaml"
@@ -361,10 +357,7 @@ func TestApplyExpansion(t *testing.T) {
 }
 
 func TestOverriddenKubeconfigPathResolution(t *testing.T) {
-	os.Setenv("SUBPATH", "test")
-	t.Cleanup(func() {
-		os.Unsetenv("SUBPATH")
-	})
+	t.Setenv("SUBPATH", "test")
 	stepRelativePath := &Step{Dir: "test_data/kubeconfig_path_resolution/"}
 	err := stepRelativePath.LoadYAML(kfile.Parse("test_data/kubeconfig_path_resolution/00-step1.yaml"))
 	assert.NoError(t, err)
@@ -426,7 +419,7 @@ func TestStepFailure(t *testing.T) {
 
 	namespace := fmt.Sprintf("kuttl-test-%s", petname.Generate(2, "-"))
 
-	err := testenv.Client.Create(context.TODO(), kubernetes.NewResource("v1", "Namespace", namespace, ""))
+	err := testenv.Client.Create(t.Context(), kubernetes.NewResource("v1", "Namespace", namespace, ""))
 	if !k8serrors.IsAlreadyExists(err) {
 		// we are ignoring already exists here because in tests we by default use retry client so this can happen
 		assert.Nil(t, err)
