@@ -2,8 +2,6 @@ package utils
 
 import (
 	"bytes"
-	"context"
-	"os"
 	"strings"
 	"testing"
 
@@ -173,15 +171,10 @@ func TestGetKubectlArgs(t *testing.T) {
 		t.Run(test.testName, func(t *testing.T) {
 			if test.env != nil || len(test.env) > 0 {
 				for key, value := range test.env {
-					os.Setenv(key, value)
+					t.Setenv(key, value)
 				}
-				defer func() {
-					for key := range test.env {
-						os.Unsetenv(key)
-					}
-				}()
 			}
-			cmd, err := GetArgs(context.TODO(), harness.Command{
+			cmd, err := GetArgs(t.Context(), harness.Command{
 				Command:    test.args,
 				Namespaced: true,
 			}, test.namespace, nil)
@@ -241,7 +234,7 @@ func TestRunScript(t *testing.T) {
 
 			logger := NewTestLogger(t, "")
 			// script runs with output
-			_, err := RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+			_, err := RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 
 			if tt.wantedErr {
 				assert.Error(t, err)
@@ -266,7 +259,7 @@ func TestRunCommand(t *testing.T) {
 
 	logger := NewTestLogger(t, "")
 	// assert foreground cmd returns nil
-	cmd, err := RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err := RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.NoError(t, err)
 	assert.Nil(t, cmd)
 	// foreground processes should have stdout
@@ -276,18 +269,21 @@ func TestRunCommand(t *testing.T) {
 	stdout = &bytes.Buffer{}
 
 	// assert background cmd returns process
-	cmd, err = RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err = RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.NoError(t, err)
 	assert.NotNil(t, cmd)
-	// no stdout for background processes
-	assert.Empty(t, strings.TrimSpace(stdout.String()))
+
+	// Wait for background process to complete before checking stdout
+	err = cmd.Wait()
+	assert.NoError(t, err)
+	assert.Equal(t, "hello", strings.TrimSpace(stdout.String()))
 
 	stdout = &bytes.Buffer{}
 	hcmd.Background = false
 	hcmd.Command = "sleep 42"
 
 	// assert foreground cmd times out
-	cmd, err = RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 2, "")
+	cmd, err = RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 2, "")
 	assert.Error(t, err)
 	assert.True(t, strings.Contains(err.Error(), "timeout"))
 	assert.Nil(t, cmd)
@@ -298,7 +294,7 @@ func TestRunCommand(t *testing.T) {
 	hcmd.Timeout = 2
 
 	// assert foreground cmd times out with command timeout
-	cmd, err = RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err = RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.Error(t, err)
 	assert.True(t, strings.Contains(err.Error(), "timeout"))
 	assert.Nil(t, cmd)
@@ -314,12 +310,12 @@ func TestRunCommandIgnoreErrors(t *testing.T) {
 
 	logger := NewTestLogger(t, "")
 	// assert foreground cmd returns nil
-	cmd, err := RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err := RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.NoError(t, err)
 	assert.Nil(t, cmd)
 
 	hcmd.IgnoreFailure = false
-	cmd, err = RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err = RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.Error(t, err)
 	assert.Nil(t, cmd)
 
@@ -328,7 +324,7 @@ func TestRunCommandIgnoreErrors(t *testing.T) {
 		Command:       "bad-command",
 		IgnoreFailure: true,
 	}
-	cmd, err = RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err = RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.Error(t, err)
 	assert.Nil(t, cmd)
 }
@@ -342,7 +338,7 @@ func TestRunCommandSkipLogOutput(t *testing.T) {
 
 	logger := NewTestLogger(t, "")
 	// test there is a stdout
-	cmd, err := RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err := RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.NoError(t, err)
 	assert.Nil(t, cmd)
 	assert.True(t, stdout.Len() > 0)
@@ -351,7 +347,7 @@ func TestRunCommandSkipLogOutput(t *testing.T) {
 	stdout = &bytes.Buffer{}
 	stderr = &bytes.Buffer{}
 	// test there is no stdout
-	cmd, err = RunCommand(context.TODO(), "", hcmd, "", stdout, stderr, logger, 0, "")
+	cmd, err = RunCommand(t.Context(), "", hcmd, "", stdout, stderr, logger, 0, "")
 	assert.NoError(t, err)
 	assert.Nil(t, cmd)
 	assert.True(t, stdout.Len() == 0)
